requirements.txt
asgiref==3.2.10
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.16.13
botocore==1.19.13
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.0
pkg_resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.11
uWSGI==2.0.20
zipp==3.4.0
